[SDAO-98] Fix and extend the Multisig test suite
Problem: The test suite was broken (due to incorrent configuration),
and not exhaustive (it lacked RotateKeys tests and duplicate keys
tests). Additionally, it used truffle, which meaned it tried to
redeploy all the contracts before the test.

Solution: Refactor the test suite, add the missing tests, use
OpenZeppelin instead of truffle to run the tests.