[SDAO-126] Deduce fee for burning tokens
Problem: Currently, if there are multiple sources that lock funds
on the same tender address, an older source may unlock the funds
of a newer source. It is important because if some person's key is
compromised and it was used in burn operations, the new owner of
the key can withdraw funds from the tender address.

Solution:
1. Track burns in Registry
2. Add fees for each address who sent tokens to tender address
3. Add tests for this logic