[SDAO-98] Remove fees from Orchestrator contract
Problem: During the "fees rehaul" a new function was added to the
Orchestrator to set the fees. However, it turned out that this
contract doesn't need to know about the fees, instead we decided
to offload fees computation to Registry. The function and the
corresponding field were never removed though.

Solution: Remove the unneeded setter and the field.