[SDAO-127] Switch to two-step ownership transfer
Problem: Currently, it is possible to invoke
`{blend,orchestrator,registry}.transferOwnership(newOwner)`
from the current owner's address to set the new owner. However,
mistakes happen, and it is possible that the new address would
have no private key associated with it. This can potentially
block the backend key rotation for all the contracts, and it
would be impossible to call `orchestrator.collectBlend`

Solution: Use two-step ownership trasfer, i.e. add a contract
that, unlike the standard Ownable, implements an additional
`acceptOwnership` method that actually transfers the ownership.
`transferOwnership` endpoint, in turn, would not actually do the
transfer but rather set a `pendingOwner` of the contract. Until
`acceptTransfer` is called, the current `owner` possesses the
ownership rights.